Improves tinker experience in Lambda environment
Enhances the tinker experience by displaying a startup message
that informs the user they are running tinker in AWS Lambda and
shows the target Lambda function name.

Also sets the raw output stream to the Lambda shell to allow for
formatted output. Updates the README with more realistic tinker examples.

Adds the `SLS_PLATFORM` env variable to the phpunit config.